#17 Moved ModelLoader to top-level package; differentiating root and regular nodes, and tokens in language specification.
