+++ cml-compiler/cml-language/src/main/java/cml/language/ModelSynthesizer.java
@@ -1,10 +1,11 @@
-package cml.language.grammar;
+package cml.language;
 
 import cml.language.features.Concept;
 import cml.language.features.Model;
 import cml.language.foundation.Property;
 import cml.language.features.Target;
 import cml.language.foundation.Type;
+import cml.language.grammar.CMLBaseListener;
 import cml.language.grammar.CMLParser.*;
 
 public class ModelSynthesizer extends CMLBaseListener
